refactor: yearn deposit split
