+++ src/features/defi/providers/yearn/components/YearnManager/Deposit/components/Status.tsx
@@ -1,0 +1,183 @@
+import { ArrowForwardIcon, CheckIcon, CloseIcon } from '@chakra-ui/icons'
+import { Box, Link, Stack, Tag } from '@chakra-ui/react'
+import { AssetNamespace, AssetReference, caip19 } from '@shapeshiftoss/caip'
+import { NetworkTypes } from '@shapeshiftoss/types'
+import { TxStatus } from 'features/defi/components/TxStatus/TxStatus'
+import { DefiParams, DefiQueryParams } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
+import { useContext } from 'react'
+import { Amount } from 'components/Amount/Amount'
+import { MiddleEllipsis } from 'components/MiddleEllipsis/MiddleEllipsis'
+import { StatusTextEnum } from 'components/RouteSteps/RouteSteps'
+import { Row } from 'components/Row/Row'
+import { Text } from 'components/Text'
+import { useBrowserRouter } from 'hooks/useBrowserRouter/useBrowserRouter'
+import { bnOrZero } from 'lib/bignumber/bignumber'
+import { marketApi } from 'state/slices/marketDataSlice/marketDataSlice'
+import { selectAssetByCAIP19, selectMarketDataById } from 'state/slices/selectors'
+import { useAppDispatch, useAppSelector } from 'state/store'
+
+import { DepositContext } from '../DepositContext'
+
+export const Status = () => {
+  const { state } = useContext(DepositContext)
+  const appDispatch = useAppDispatch()
+  const { query, history: browserHistory } = useBrowserRouter<DefiQueryParams, DefiParams>()
+  const { chain, contractAddress: vaultAddress, tokenId } = query
+
+  const network = NetworkTypes.MAINNET
+  const assetNamespace = AssetNamespace.ERC20
+  const assetCAIP19 = caip19.toCAIP19({ chain, network, assetNamespace, assetReference: tokenId })
+  const feeAssetCAIP19 = caip19.toCAIP19({
+    chain,
+    network,
+    assetNamespace: AssetNamespace.Slip44,
+    assetReference: AssetReference.Ethereum,
+  })
+  const asset = useAppSelector(state => selectAssetByCAIP19(state, assetCAIP19))
+  const marketData = useAppSelector(state => selectMarketDataById(state, assetCAIP19))
+  if (!marketData) appDispatch(marketApi.endpoints.findByCaip19.initiate(assetCAIP19))
+  const feeAsset = useAppSelector(state => selectAssetByCAIP19(state, feeAssetCAIP19))
+  const feeMarketData = useAppSelector(state => selectMarketDataById(state, feeAssetCAIP19))
+  const vaultCAIP19 = caip19.toCAIP19({
+    chain,
+    network,
+    assetNamespace,
+    assetReference: vaultAddress,
+  })
+  const vaultAsset = useAppSelector(state => selectAssetByCAIP19(state, vaultCAIP19))
+
+  const handleViewPosition = () => {
+    browserHistory.push('/defi')
+  }
+
+  const handleCancel = () => {
+    browserHistory.goBack()
+  }
+
+  if (!state) return null
+
+  const apy = state.vault.metadata?.apy?.net_apy
+  const annualYieldCrypto = bnOrZero(state.deposit?.cryptoAmount).times(bnOrZero(apy))
+  const annualYieldFiat = annualYieldCrypto.times(marketData.price)
+
+  let statusIcon: React.ReactElement = <ArrowForwardIcon />
+  let statusText = StatusTextEnum.pending
+  if (state.deposit.txStatus === 'success') {
+    statusText = StatusTextEnum.success
+    statusIcon = <CheckIcon color='green' />
+  }
+  if (state.deposit.txStatus === 'failed') {
+    statusText = StatusTextEnum.failed
+    statusIcon = <CloseIcon color='red' />
+  }
+
+  return (
+    <TxStatus
+      onClose={handleCancel}
+      onContinue={state.deposit.txStatus === 'success' ? handleViewPosition : undefined}
+      loading={state.loading}
+      statusText={statusText}
+      statusIcon={statusIcon}
+      continueText='modals.status.position'
+      closeText='modals.status.close'
+      assets={[
+        {
+          ...asset,
+          cryptoAmount: state.deposit.cryptoAmount,
+          fiatAmount: state.deposit.fiatAmount,
+        },
+        {
+          ...vaultAsset,
+          cryptoAmount: bnOrZero(state.deposit.cryptoAmount)
+            .div(bnOrZero(state.pricePerShare).div(`1e+${state.vault.decimals}`))
+            .toString(),
+          fiatAmount: state.deposit.fiatAmount,
+        },
+      ]}
+    >
+      <Stack spacing={4}>
+        <Row>
+          <Row.Label>
+            <Text translation='modals.status.transactionId' />
+          </Row.Label>
+          <Row.Value>
+            <Link
+              href={`${asset.explorerTxLink}/${state.txid}`}
+              isExternal
+              color='blue.500'
+              fontWeight='bold'
+            >
+              <MiddleEllipsis address={state.txid || ''} />
+            </Link>
+          </Row.Value>
+        </Row>
+        <Row>
+          <Row.Label>
+            <Text translation='modals.confirm.depositTo' />
+          </Row.Label>
+          <Row.Value fontWeight='bold'>Yearn Finance</Row.Value>
+        </Row>
+        <Row>
+          <Row.Label>
+            <Text
+              translation={
+                state.deposit.txStatus === 'pending'
+                  ? 'modals.status.estimatedGas'
+                  : 'modals.status.gasUsed'
+              }
+            />
+          </Row.Label>
+          <Row.Value>
+            <Box textAlign='right'>
+              <Amount.Fiat
+                fontWeight='bold'
+                value={bnOrZero(
+                  state.deposit.txStatus === 'pending'
+                    ? state.deposit.estimatedGasCrypto
+                    : state.deposit.usedGasFee,
+                )
+                  .div(`1e+${feeAsset.precision}`)
+                  .times(feeMarketData.price)
+                  .toFixed(2)}
+              />
+              <Amount.Crypto
+                color='gray.500'
+                value={bnOrZero(
+                  state.deposit.txStatus === 'pending'
+                    ? state.deposit.estimatedGasCrypto
+                    : state.deposit.usedGasFee,
+                )
+                  .div(`1e+${feeAsset.precision}`)
+                  .toFixed(5)}
+                symbol='ETH'
+              />
+            </Box>
+          </Row.Value>
+        </Row>
+        <Row>
+          <Row.Label>
+            <Text translation='modals.confirm.deposit.averageApr' />
+          </Row.Label>
+          <Tag colorScheme='green'>
+            <Amount.Percent value={String(apy)} />
+          </Tag>
+        </Row>
+        <Row>
+          <Row.Label>
+            <Text translation='modals.confirm.deposit.estimatedReturns' />
+          </Row.Label>
+          <Row.Value>
+            <Box textAlign='right'>
+              <Amount.Fiat fontWeight='bold' value={annualYieldFiat.toFixed(2)} />
+              <Amount.Crypto
+                color='gray.500'
+                value={annualYieldCrypto.toFixed(5)}
+                symbol={asset.symbol}
+              />
+            </Box>
+          </Row.Value>
+        </Row>
+      </Stack>
+    </TxStatus>
+  )
+}
